Move EndpointInfo and friends to iroh-dns
