+++ iroh-dns-server/examples/resolve.rs
@@ -2,7 +2,7 @@ use clap::{Parser, ValueEnum};
 use iroh::{
     EndpointId,
     address_lookup::dns::{N0_DNS_ENDPOINT_ORIGIN_PROD, N0_DNS_ENDPOINT_ORIGIN_STAGING},
-    dns::DnsResolver,
+    dns::{DnsResolver, DnsResolverExt},
 };
 use n0_error::{Result, StackResultExt, StdResultExt};
