+++ iroh-dns/build.rs
@@ -1,0 +1,8 @@
+use cfg_aliases::cfg_aliases;
+
+fn main() {
+    cfg_aliases! {
+        wasm_browser: { all(target_family = "wasm", target_os = "unknown") },
+        with_crypto_provider: { any(feature = "tls-ring", feature = "tls-aws-lc-rs") }
+    }
+}
